PB1557+PB1558 (w59k): ALPHABET full-set ordinals, LOCALE bind order
Cluster 2 of train 62 (branch worktree-agent-abb030ce461bd019f, base
cdb084309).

PB1557: an ALPHABET naming all 65,536 native characters (legal: §12.3.7.3
SR14 b4/c4 "shall not exceed") wrapped its 16-bit position count to 0,
emptying the table's specified block — CHAR/ORD gave wrong answers
(RT66=65504) and CLASS/SYMBOLIC ... IN refused every ordinal. The count is
`int` in AlphabetLiteralPhrase, ImplementorCodeNames.SingleByteCodePage and
CollatingTable.Build (which now asserts its invariants). Sibling: the
SPECIAL-NAMES literal-phrase ordinals read through int.TryParse, so an
11-digit ordinal was reported under the noninteger class rule; one reader
(IntegerLiteralDigits + saturating OrdinalValue) works from the parse.

PB1558: ALPHABET ... IS LOCALE locale-name-2 was resolved in source order
although §12.3.7.2 / §5.2.1 put the LOCALE clause after it;
SwitchBindSpecialNames binds in dependency order (declaring clauses, then
ALPHABET, then CLASS/SYMBOLIC/OBJECT-COMPUTER). Eight goldens that pinned
the non-conforming LOCALE-before-ALPHABET order were rewritten to the
format's order (pb64t1_* x6, 2014/pb764_format_keywords_not_names,
negative pb64t1-alphabet-locale-name-undeclared).

Lander conflict resolution (DataBinder.Switches.cs, SwitchBindSymbolic):
main's train 60 (PB1058) had already routed SYMBOLIC CHARACTERS integer-1
through IntegerOperandRules.HostValue, THE ONE integer-n reader; this
cluster had introduced OrdinalValue there. Resolved to HostValue (it
saturates identically, and the operand IS an integer-n), keeping the
cluster's ordText for the SR16 message; OrdinalValue now serves the
literal phrases only and its doc comment says so. The negative
pb1557-symbolic-ordinal-too-long passes by name (COBOLNET1670).

Goldens: 85/pb1557_full_alphabet_ordinals_85,
2002/l1c30_full_alphabet_in_ordinals,
2002/l1c30_alphabet_named_locale_format_order; negative
pb1557-symbolic-ordinal-too-long. Tests: SpecialNamesClauseOrderTests,
AlphabetLiteralPhraseTests, ClassClauseForPhraseTests,
CodedCharacterSetTests.

Rows: w59k-batch.json, 5 witness records (+14 test-ref, +5
code-location), GAP 1256 -> 1254 (+2 on the merged tree; golden batch 2
had already closed two). PB1557, PB1558 landed.

+++ src/Cobol.Net.Compiler/Binding/ImplementorCodeNames.cs
@@ -158,7 +158,7 @@ public static class ImplementorCodeNames
     {
         var page = Encoding.GetEncoding(codePage);
         char[] toNative = page.GetChars([.. Enumerable.Range(0, 256).Select(b => (byte)b)]);
-        var pos = new Dictionary<char, ushort>(256);
+        var pos = new Dictionary<char, int>(256);
         var order = new List<char>(256);
         var repByPos = new List<char>(256);
         for (int unit = 0; unit < toNative.Length; unit++)
@@ -166,7 +166,7 @@ public static class ImplementorCodeNames
             // A registered single-byte page is a bijection, so no code unit can collide; the guard states the
             // requirement rather than trusting it (a page that is not one cannot define a collating sequence,
             // because two native characters would share a position without an ALSO phrase saying so).
-            if (!pos.TryAdd(toNative[unit], (ushort)unit))
+            if (!pos.TryAdd(toNative[unit], unit))
                 throw new InvalidOperationException(
                     $"code page {codePage} maps two code units to U+{(int)toNative[unit]:X4}: it cannot define "
                     + "a collating sequence (ISO §12.3.7.4 GR7 i)");
@@ -174,6 +174,6 @@ public static class ImplementorCodeNames
             repByPos.Add(toNative[unit]);
         }
         return new ImplementorCodeName(name, National: false, toNative.Length,
-            CollatingTable.Build(pos, order, repByPos, (ushort)toNative.Length, national: false), toNative);
+            CollatingTable.Build(pos, order, repByPos, toNative.Length, national: false), toNative);
     }
 }
